+++ lib/hooks/interface.go
@@ -1,0 +1,6 @@
+package hooks
+
+type HookType interface {
+	Cmd() string
+	Env() []string
+}